fix test stuff
