+++ Workflow.Site/packages.config
@@ -45,7 +45,7 @@
   <package id="System.Collections.Immutable" version="1.1.36" targetFramework="net46" />
   <package id="System.Reflection.Metadata" version="1.0.21" targetFramework="net46" />
   <package id="System.Threading.Tasks.Dataflow" version="4.7.0" targetFramework="net46" />
-  <package id="System.ValueTuple" version="4.4.0" targetFramework="net46" />
+  <package id="System.ValueTuple" version="4.5.0" targetFramework="net46" />
   <package id="Umbraco.ModelsBuilder" version="3.0.10" targetFramework="net46" />
   <package id="UmbracoCms" version="7.10.4" targetFramework="net46" />
   <package id="UmbracoCms.Core" version="7.10.4" targetFramework="net46" />
