+++ Workflow.Tests/packages.config
@@ -1,14 +1,14 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
   <package id="AutoMapper" version="3.3.1" targetFramework="net461" />
-  <package id="Chauffeur" version="1.1.0-v11-142" targetFramework="net461" />
-  <package id="Chauffeur.TestingTools" version="1.1.0-v11-142" targetFramework="net461" />
+  <package id="Chauffeur" version="1.4.0" targetFramework="net462" />
+  <package id="Chauffeur.TestingTools" version="1.4.0" targetFramework="net462" />
   <package id="ClientDependency" version="1.9.6" targetFramework="net461" />
   <package id="ClientDependency-Mvc" version="1.8.0.0" targetFramework="net461" />
   <package id="ClientDependency-Mvc5" version="1.8.0.0" targetFramework="net461" />
   <package id="DotNetSeleniumExtras.WaitHelpers" version="3.11.0" targetFramework="net462" />
   <package id="Examine" version="0.1.89" targetFramework="net461" />
-  <package id="FSharp.Core" version="4.3.4" targetFramework="net461" />
+  <package id="FSharp.Core" version="4.5.4" targetFramework="net462" />
   <package id="GDev.Umbraco.Testing" version="2.0.0" targetFramework="net462" />
   <package id="HtmlAgilityPack" version="1.4.9.5" targetFramework="net461" />
   <package id="ImageProcessor" version="2.5.6" targetFramework="net461" />
@@ -28,8 +28,10 @@
   <package id="Microsoft.AspNet.WebApi.Core" version="5.2.3" targetFramework="net461" />
   <package id="Microsoft.AspNet.WebApi.WebHost" version="5.2.3" targetFramework="net461" />
   <package id="Microsoft.AspNet.WebPages" version="3.2.3" targetFramework="net461" />
+  <package id="Microsoft.CodeCoverage" version="15.9.0" targetFramework="net462" />
   <package id="Microsoft.IO.RecyclableMemoryStream" version="1.2.2" targetFramework="net461" />
   <package id="Microsoft.Net.Http" version="2.0.20710.0" targetFramework="net461" />
+  <package id="Microsoft.NET.Test.Sdk" version="15.9.0" targetFramework="net462" />
   <package id="Microsoft.Owin" version="3.1.0" targetFramework="net461" />
   <package id="Microsoft.Owin.Host.SystemWeb" version="3.1.0" targetFramework="net461" />
   <package id="Microsoft.Owin.Security" version="3.1.0" targetFramework="net461" />
@@ -47,7 +49,7 @@
   <package id="SharpZipLib" version="0.86.0" targetFramework="net461" />
   <package id="System.IO.Abstractions" version="1.4.0.93" targetFramework="net461" />
   <package id="System.Threading.Tasks.Dataflow" version="4.7.0" targetFramework="net461" />
-  <package id="System.ValueTuple" version="4.4.0" targetFramework="net461" />
+  <package id="System.ValueTuple" version="4.5.0" targetFramework="net462" />
   <package id="UmbracoCms.Core" version="7.10.4" targetFramework="net462" />
   <package id="xunit" version="2.3.1" targetFramework="net461" />
   <package id="xunit.abstractions" version="2.0.1" targetFramework="net461" />
@@ -56,4 +58,5 @@
   <package id="xunit.core" version="2.3.1" targetFramework="net461" />
   <package id="xunit.extensibility.core" version="2.3.1" targetFramework="net461" />
   <package id="xunit.extensibility.execution" version="2.3.1" targetFramework="net461" />
+  <package id="xunit.runner.visualstudio" version="2.4.1" targetFramework="net462" developmentDependency="true" />
 </packages>
