+++ Workflow/packages.config
@@ -37,7 +37,7 @@
   <package id="System.Net.Http" version="4.0.0" targetFramework="net45" />
   <package id="System.Net.Http.Formatting.Extension" version="5.2.3.0" targetFramework="net45" />
   <package id="System.Threading.Tasks.Dataflow" version="4.7.0" targetFramework="net46" />
-  <package id="System.ValueTuple" version="4.4.0" targetFramework="net46" />
+  <package id="System.ValueTuple" version="4.5.0" targetFramework="net46" />
   <package id="UmbracoCms.Core" version="7.10.4" targetFramework="net46" />
   <package id="UrlRewritingNet" version="2.0.7" targetFramework="net45" />
   <package id="xmlrpcnet" version="2.5.0" targetFramework="net45" />
